test(workbench): move page objects copied from @scion/toolkit to @scion folder
