+++ projects/scion/e2e-testing/src/@scion/components.internal/README.txt
@@ -1,0 +1,5 @@
+Copied from @scion/components.internal because we cannot reference ESM modules in Playwright.
+
+---
+Source: https://github.com/SchweizerischeBundesbahnen/scion-toolkit
+---
