+++ projects/scion/e2e-testing/src/@scion/components.internal/tabbar.po/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2018-2022 Swiss Federal Railways
+ *
+ * This program and the accompanying materials are made
+ * available under the terms from the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ *  SPDX-License-Identifier: EPL-2.0
+ */
+
+export * from './tabbar.po';
