+++ projects/scion/e2e-testing/src/workbench-client/page-object/popup-opener-page.po.ts
@@ -13,9 +13,9 @@ import {AppPO} from '../../app.po';
 import {ViewPO} from '../../view.po';
 import {Qualifier} from '@scion/microfrontend-platform';
 import {BottomLeftPoint, BottomRightPoint, PopupOrigin, TopLeftPoint, TopRightPoint} from '@scion/workbench';
-import {SciParamsEnterPO} from '../../components.internal/params-enter.po';
-import {SciAccordionPO} from '../../components.internal/accordion.po';
-import {SciCheckboxPO} from '../../components.internal/checkbox.po';
+import {SciParamsEnterPO} from '../../@scion/components.internal/params-enter.po';
+import {SciAccordionPO} from '../../@scion/components.internal/accordion.po';
+import {SciCheckboxPO} from '../../@scion/components.internal/checkbox.po';
 import {Locator} from '@playwright/test';
 import {ElementSelectors} from '../../helper/element-selectors';
 import {SciRouterOutletPO} from './sci-router-outlet.po';
