+++ projects/scion/e2e-testing/src/workbench-client/page-object/register-workbench-intention-page.po.ts
@@ -11,7 +11,7 @@
 import {AppPO} from '../../app.po';
 import {ViewTabPO} from '../../view-tab.po';
 import {Intention, Qualifier} from '@scion/microfrontend-platform';
-import {SciParamsEnterPO} from '../../components.internal/params-enter.po';
+import {SciParamsEnterPO} from '../../@scion/components.internal/params-enter.po';
 import {Locator} from '@playwright/test';
 import {ElementSelectors} from '../../helper/element-selectors';
