+++ projects/scion/e2e-testing/src/workbench-client/page-object/view-page.po.ts
@@ -13,10 +13,10 @@ import {AppPO} from '../../app.po';
 import {ViewPO} from '../../view.po';
 import {Params} from '@angular/router';
 import {WorkbenchViewCapability} from '@scion/workbench-client';
-import {SciAccordionPO} from '../../components.internal/accordion.po';
-import {SciPropertyPO} from '../../components.internal/property.po';
-import {SciCheckboxPO} from '../../components.internal/checkbox.po';
-import {SciParamsEnterPO} from '../../components.internal/params-enter.po';
+import {SciAccordionPO} from '../../@scion/components.internal/accordion.po';
+import {SciPropertyPO} from '../../@scion/components.internal/property.po';
+import {SciCheckboxPO} from '../../@scion/components.internal/checkbox.po';
+import {SciParamsEnterPO} from '../../@scion/components.internal/params-enter.po';
 import {Locator} from '@playwright/test';
 import {ElementSelectors} from '../../helper/element-selectors';
 import {SciRouterOutletPO} from './sci-router-outlet.po';
